Create Modal component and use it with CreateTargetForm
