+++ src/components/common/Modal/styles.js
